Improve permissions handling and remove unused code
