+++ lib/android_permission_status.dart
@@ -1,0 +1,16 @@
+///
+///  Copyright © 2019 PSPDFKit GmbH. All rights reserved.
+///
+///  THIS SOURCE CODE AND ANY ACCOMPANYING DOCUMENTATION ARE PROTECTED BY INTERNATIONAL COPYRIGHT LAW
+///  AND MAY NOT BE RESOLD OR REDISTRIBUTED. USAGE IS BOUND TO THE PSPDFKIT LICENSE AGREEMENT.
+///  UNAUTHORIZED REPRODUCTION OR DISTRIBUTION IS SUBJECT TO CIVIL AND CRIMINAL PENALTIES.
+///  This notice may not be removed from this file.
+///
+part of pspdfkit;
+
+enum AndroidPermissionStatus {
+  notDetermined,
+  denied,
+  authorized,
+  deniedNeverAsk
+}
